Rename chains module to chain
