+++ packages/panna-sdk/src/core/client/README.md
@@ -210,6 +210,6 @@ Now that you have a configured client, explore these SDK modules:
 
 - **[Wallet](../wallet/README.md)** - User authentication and account management
 - **[Transactions](../transaction/README.md)** - Send and prepare blockchain transactions
-- **[Chains](../chains/README.md)** - Configure supported blockchain networks
+- **[Chain](../chain/README.md)** - Configure supported blockchain networks
 - **[Onramp](../onramp/README.md)** - Integrate fiat-to-crypto payments
 - **[Utils](../utils/README.md)** - Helper functions for common operations
