+++ packages/panna-sdk/src/core/index.ts
@@ -1,5 +1,5 @@
 export * from './client';
-export * from './chains';
+export * from './chain';
 export * from './wallet';
 export * from './utils';
 export * from './defaults';
