+++ packages/panna-sdk/src/core/onramp/README.md
@@ -228,4 +228,4 @@ const provider = getBestProvider(countryCode);
 - Explore [Wallet Module](../wallet/README.md) for user authentication and account setup
 - Review [Transaction Module](../transaction/README.md) for handling purchased tokens
 - Check [Utils Module](../utils/README.md) for balance tracking and fiat conversions
-- Learn about [Chains Module](../chains/README.md) for multi-chain onramp support
+- Learn about [Chain Module](../chain/README.md) for multi-chain onramp support
