+++ packages/panna-sdk/src/core/onramp/onramp.ts
@@ -1,6 +1,6 @@
 import { getCountry } from 'iso-3166-1-alpha-2';
 import { Bridge } from 'thirdweb';
-import { lisk } from '../chains';
+import { lisk } from '../chain';
 import type { TokenFiatPrice } from '../utils/types';
 import { COUNTRY_PROVIDER_MAP, PROVIDERS } from './constants';
 import type {
