+++ packages/panna-sdk/src/core/transaction/transaction.test.ts
@@ -1,5 +1,5 @@
 import * as thirdweb from 'thirdweb';
-import type { Chain } from '../chains/types';
+import type { Chain } from '../chain/types';
 import type { PannaClient } from '../client';
 import type { Address } from '../types/external';
 import {
